Remove aliases in the author format
We do not use the alias anywhere in the plugin so we should really
remove it.

There is more reason to remove it since we can use built-in functions of
git to generate the co-author list that makes it simple to list in the
complete menu and also appending to the co-author format.

We update the README in this change as well to reflect the changes and
also provide git examples that we can use to generate a list from a
project.